presentation/src/main/java/com/shareit/presentation/UserController.java
package com.shareit.presentation;

import com.shareit.domain.dto.CreateUser;
import com.shareit.domain.dto.User;
import com.shareit.domain.dto.UserCreated;
import com.shareit.service.UserService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;


@RestController
@RequestMapping("/v1/user")
public class UserController {
    private UserService userService;

    public UserController(UserService userService) {
        this.userService = userService;
    }

    @GetMapping("/{id}")
    public ResponseEntity<User> getUserById(@PathVariable("id") Long userId) {
        //TODO Return ResponseEntity or create a exception handler to handle different type of exceptions
        return ResponseEntity.ok(userService.findById(userId));
    }

    @PostMapping
    public ResponseEntity<UserCreated> userRegister(@Valid @RequestBody CreateUser createUser) {
        return ResponseEntity.ok(new UserCreated(userService.createUser(createUser)));
    }
}
